updated _fields with suggestion from stack overflow. It now works

+++ theNetowrkSocketsAndSniffing/sniffer_ip_header_decode.py
@@ -20,8 +20,8 @@ class IP(Structure):
         ("ttl", c_ubyte),
         ("protocol_num", c_ubyte),
         ("sum", c_ushort),
-        ("src", c_ulong),
-        ("dst", c_ulong)
+        ("src", c_uint32),
+        ("dst", c_uint32)
     ]
 
     def __new__(self, socket_buffer=None):
@@ -32,8 +32,8 @@ class IP(Structure):
         self.protocol_map = {1:"ICMP", 6:"TCP", 17:"UDP"}
 
         #human readable IP addresses
-        self.src_address = socket.inet_ntoa(struct.pack("<L",self.src))
-        self.dst_address = socket.inet_ntoa(struct.pack("<L",self.dst))
+        self.src_address = socket.inet_ntoa(struct.pack('<L',self.src))
+        self.dst_address = socket.inet_ntoa(struct.pack('<L',self.dst))
 
         #human readable protocol
         try:
